Move qt_interop out of src directory and use fully qualified #includes.
