+++ qt_interop/qimage.h
