+++ qt_interop/vecmath.h
